% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boys}
\alias{boys}
\title{Boys.}
\format{
A data frame with one variable \code{Voornaam}, the first name.
}
\usage{
boys
}
\description{
Collection of Dutch names for boys.
}
\keyword{datasets}
