% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\alias{circle_circum}
\alias{circle_area}
\title{Circle Circumference and Area calculation}
\usage{
circle_circum(r)

circle_area(r)
}
\arguments{
\item{r}{numeric}
}
\value{
numeric
}
\description{
\code{circle_circum()} and \code{circle_area()} Respectively the calculation of
the circumference and the area of a circle with radius r.
}
\section{Functions}{
\itemize{
\item \code{circle_circum()}: circumference calculation

\item \code{circle_area()}: area calculation

}}
\examples{
circle_circum(12)
circle_area(12.3)

}
