% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmonths.R
\docType{data}
\name{Dutch_Month_Names}
\alias{Dutch_Month_Names}
\alias{month.abb.nl}
\alias{month.name.nl}
\title{Dutch names for the months of the year.}
\format{
An object of class \code{character} of length 12.

An object of class \code{character} of length 12.
}
\usage{
month.abb.nl

month.name.nl
}
\value{
character
}
\description{
\code{month.abb.nl} three-letter abbreviations
\code{month.name.nl} full names
}
\section{Functions}{
\itemize{
\item \code{month.abb.nl}: three-letter abbreviations

\item \code{month.name.nl}: full names

}}
\examples{
month.abb.nl
month.abb.nl[1:3]
month.name.nl
month.name.nl[4:6]

}
\keyword{datasets}
