% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{girls}
\alias{girls}
\title{Girls.}
\format{
A data frame with one variable \code{Voornaam}, the first name.
}
\usage{
girls
}
\description{
Collection of Dutch names for girls
}
\keyword{datasets}
