% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jongens}
\alias{jongens}
\title{Jongens.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 199 rows and 1 columns.
}
\usage{
jongens
}
\description{
Verzameling van voornamen van jongens
}
\keyword{datasets}
