% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meisjes}
\alias{meisjes}
\title{Meisjes.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 199 rows and 1 columns.
}
\usage{
meisjes
}
\description{
Verzameling van voornamen van meisjes.
}
\keyword{datasets}
