% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people}
\alias{people}
\title{Colored People.}
\format{
A data frame with three variables \code{Voornaam}, \code{Achternaam}
en \code{Geslacht}. Respectively the first name, last name and gender.
}
\usage{
people
}
\description{
Collection of Dutch names for people where the names
are made from names of colors.
}
\keyword{datasets}
